moved functions to service
