+++ src/main/app/src/app/admin/smart-documents.service.ts
@@ -9,6 +9,28 @@ import { FoutAfhandelingService } from "../fout-afhandeling/fout-afhandeling.ser
 import { ZacHttpClient } from "../shared/http/zac-http-client";
 import { GeneratedType } from "../shared/utils/generated-types";
 
+export type MappedSmartDocumentsTemplateWithParentId =
+  GeneratedType<"RestMappedSmartDocumentsTemplate"> & {
+    parentGroupId?: string | null;
+  };
+export type MappedSmartDocumentsTemplateGroupWithParentId = Omit<
+  GeneratedType<"RestMappedSmartDocumentsTemplateGroup">,
+  "groups" | "templates"
+> & {
+  groups?: MappedSmartDocumentsTemplateGroupWithParentId[] | null;
+  templates?: MappedSmartDocumentsTemplateWithParentId[] | null;
+};
+
+export type PlainTemplateMappings = Omit<
+  MappedSmartDocumentsTemplateWithParentId,
+  "name"
+>;
+
+export type MappedSmartDocumentsTemplateFlattenedGroupWithParentId = Omit<
+  MappedSmartDocumentsTemplateGroupWithParentId,
+  "groups"
+>;
+
 @Injectable({ providedIn: "root" })
 export class SmartDocumentsService {
   constructor(
@@ -142,4 +164,127 @@ export class SmartDocumentsService {
 
     return result;
   }
+
+  getMappedTemplates = (
+    data: MappedSmartDocumentsTemplateGroupWithParentId[],
+  ): GeneratedType<"RestMappedSmartDocumentsTemplateGroup">[] => {
+    return data
+      .map((group) => {
+        const templates = (group.templates || [])
+          .filter((template) => template.informatieObjectTypeUUID)
+          .map(({ parentGroupId, ...template }) => template);
+
+        const groups = group.groups
+          ? this.getMappedTemplates(group.groups)
+          : [];
+
+        if (templates.length || groups.length) {
+          const { ...cleanedGroup } = group;
+          return {
+            ...cleanedGroup,
+            templates,
+            groups,
+          };
+        }
+
+        return null;
+      })
+      .filter(Boolean);
+  };
+
+  flattenNestedGroupsToRootGroups = (
+    data: MappedSmartDocumentsTemplateGroupWithParentId[],
+  ): MappedSmartDocumentsTemplateFlattenedGroupWithParentId[] => {
+    return data.flatMap((item) => {
+      const rootItemWithoutGroups = {
+        id: item.id,
+        name: item.name,
+        templates: item.templates,
+      };
+
+      const flattenedGroups =
+        item.groups?.flatMap((group) => {
+          const flattenedSubGroups = this.flattenNestedGroupsToRootGroups(
+            group.groups ?? [],
+          );
+
+          return [group, ...flattenedSubGroups];
+        }) || [];
+
+      return [rootItemWithoutGroups, ...flattenedGroups];
+    });
+  };
+
+  addTemplateMappings = (
+    groups: GeneratedType<"RestSmartDocumentsTemplateGroup">[],
+    uuidsToAdd: PlainTemplateMappings[],
+  ): MappedSmartDocumentsTemplateGroupWithParentId[] => {
+    const updateGroup = (
+      group: MappedSmartDocumentsTemplateGroupWithParentId,
+    ): MappedSmartDocumentsTemplateGroupWithParentId => ({
+      ...group,
+      templates: (group.templates || []).map((template) => ({
+        ...template,
+        informatieObjectTypeUUID: uuidsToAdd.find(
+          (uuidItem) =>
+            uuidItem.id === template.id &&
+            uuidItem.parentGroupId === template.parentGroupId,
+        )?.informatieObjectTypeUUID,
+      })),
+      groups: group.groups ? group.groups.map(updateGroup) : [],
+    });
+
+    return groups.map(updateGroup);
+  };
+
+  getTemplateMappings = (
+    data: MappedSmartDocumentsTemplateGroupWithParentId[],
+  ): PlainTemplateMappings[] => {
+    return data.flatMap((item) => {
+      const itemTemplates =
+        item.templates?.map((template) => ({
+          id: template.id,
+          parentGroupId: template.parentGroupId,
+          informatieObjectTypeUUID: template.informatieObjectTypeUUID,
+        })) ?? [];
+
+      const groupTemplates =
+        item.groups?.flatMap((group) => [
+          ...group.templates.map((template) => ({
+            id: template.id,
+            parentGroupId: template.parentGroupId,
+            informatieObjectTypeUUID: template.informatieObjectTypeUUID,
+          })),
+
+          ...this.getTemplateMappings(group.groups || []),
+        ]) ?? [];
+
+      return [...itemTemplates, ...groupTemplates];
+    });
+  };
+
+  addParentIdToTemplates = <
+    T extends
+      | GeneratedType<"RestSmartDocumentsTemplateGroup">
+      | GeneratedType<"RestMappedSmartDocumentsTemplateGroup">,
+  >(
+    data?: T[],
+  ): (Omit<T, "templates"> & {
+    templates: (T["templates"][number] & { parentGroupId: string })[];
+  })[] => {
+    if (!data) {
+      return [];
+    }
+
+    return data.map((group) => {
+      return {
+        ...group,
+        templates: group.templates?.map((template: T["templates"][number]) => ({
+          ...template,
+          parentGroupId: group.id,
+        })),
+        groups: group.groups ? this.addParentIdToTemplates(group.groups) : [],
+      };
+    });
+  };
 }
